source/python/services/discord-to-telegram-forwarder/pyproject.toml
[tool.poetry]
name = "discord-telegram-forwarder"
version = "0.1.1"
description = ""
authors = ["marklidenberg <marklidenberg@gmail.com>"]

[tool.poetry.dependencies]
python = "3.11.7"
aiohttp = "3.9.3"
cachetools = "5.2.0"
loguru = "0.6.0"
environs = "9.5.0"
python-dateutil = "2.8.2"
path = "16.4.0"
tqdm = "4.64.0"
sortedcontainers = "2.4.0"
python-telegram-bot = {version = "20.0a4", allow-prereleases = true}
Telethon = "1.24.0"
fire = "0.4.0"
anyconfig = "0.13.0"
requests = "2.28.1"
retry = "0.9.2"
pyperclip = "1.8.2"
more-itertools = "9.0.0"
"discord.py" = "2.3.1"
inline-snapshot = "0.3.2"
pytest = "7.4.0"
pyyaml = "6.0.1"
pymaybe = "0.1.6"
better-exceptions = "0.3.3"
emoji = "2.8.0"
python-box = "7.1.1"
pendulum = "3.0.0"
openai = "1.16.2"
jupyter = "1.0.0"
pydantic = "2.6.4"
pydantic-settings = "^2.2.1"
opencv-python = "^4.9.0.80"
pillow = "^10.3.0"
matplotlib = "^3.8.4"
pandas = "^2.2.2"
diskcache = "^5.6.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
